requirements.txt
--index-url https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121

numpy==1.24.4
scipy==1.10.1
pillow==9.5.0
opencv-python-headless==4.7.0.72
kornia==0.6.3
einops==0.8.1
tqdm==4.62.3
easydict==1.11
matplotlib==3.5.0
pytorch_lightning==1.9.0
gradio==4.0.0
onnxscript==0.1.0
torchmetrics==1.3.2
omegaconf==2.3.0
scikit-learn==1.3.2
scikit-image==0.21.0
prettytable==3.10.0
tabulate==0.9.0
shapely==2.0.2
psutil==5.9.8

open-clip-torch==2.0.2
ftfy
regex
huggingface-hub
safetensors

segmentation-models-pytorch @ git+https://github.com/ragavsachdeva/segmentation_models.pytorch.git@0092ee4d6f851d89a4a401bb2dfa6187660b8dd3

timm==0.4.12
